+++ ogre2/src/Ogre2RayQuery.cc
@@ -97,8 +97,27 @@ void Ogre2RayQuery::SetFromCamera(const CameraPtr &_camera,
   Ogre::Ray ray = ogre2ObjectInterface->OgreCamera()->getCameraToViewportRay(
       screenPos.X(), screenPos.Y());
 
-  this->origin = Ogre2Conversions::Convert(ray.getOrigin());
-  this->direction = Ogre2Conversions::Convert(ray.getDirection());
+  auto originMath = Ogre2Conversions::Convert(ray.getOrigin());
+  if (originMath.IsFinite())
+  {
+    this->origin = originMath;
+  }
+  else
+  {
+    ignwarn << "Attempted to set non-finite origin from camera ["
+            << camera->Name() << "]" << std::endl;
+  }
+
+  auto directionMath = Ogre2Conversions::Convert(ray.getDirection());
+  if (directionMath.IsFinite())
+  {
+    this->direction = directionMath;
+  }
+  else
+  {
+    ignwarn << "Attempted to set non-finite direction from camera ["
+            << camera->Name() << "]" << std::endl;
+  }
 
   this->dataPtr->imgPos.X() = static_cast<int>(
       screenPos.X() * _camera->ImageWidth());